requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
pyparsing==3.2.0
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
shiboken6==6.8.0.2
six==1.16.0
tzdata==2024.2
